requirements.txt
google-genai==1.39.1